+++ qupath-gui-fx/src/main/java/qupath/lib/gui/QuPathGUI.java
@@ -58,7 +58,6 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -179,13 +178,13 @@ import javafx.scene.text.TextAlignment;
 import javafx.stage.Screen;
 import javafx.stage.Stage;
 import javafx.stage.Window;
+import javafx.stage.WindowEvent;
 import javafx.util.Duration;
 import jfxtras.scene.menu.CirclePopupMenu;
 import qupath.lib.algorithms.IntensityFeaturesPlugin;
 import qupath.lib.algorithms.TilerPlugin;
 import qupath.lib.common.GeneralTools;
 import qupath.lib.common.ThreadTools;
-import qupath.lib.gui.commands.AnnotationCombineCommand;
 import qupath.lib.gui.commands.BrightnessContrastCommand;
 import qupath.lib.gui.commands.CommandListDisplayCommand;
 import qupath.lib.gui.commands.CopyViewToClipboardCommand;
@@ -292,6 +291,7 @@ import qupath.lib.gui.scripting.ScriptEditor;
 import qupath.lib.gui.tools.ColorToolsFX;
 import qupath.lib.gui.tools.CommandFinderTools;
 import qupath.lib.gui.tools.GuiTools;
+import qupath.lib.gui.tools.MenuTools;
 import qupath.lib.gui.viewer.DragDropFileImportListener;
 import qupath.lib.gui.viewer.ModeWrapper;
 import qupath.lib.gui.viewer.OverlayOptions;
@@ -600,7 +600,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		undoRedoManager = new UndoRedoManager(this);
 
 		initializingMenus.set(true);
-		menuBar.setUseSystemMenuBar(true);
+		menuBar.useSystemMenuBarProperty().bindBidirectional(PathPrefs.useSystemMenubarProperty());
 		createMenuBar();
 		pane.setTop(menuBar);
 		
@@ -785,7 +785,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		});
 
 		if (editor instanceof DefaultScriptEditor) {
-			addMenuItems(
+			MenuTools.addMenuItems(
 					menuAutomate,
 					null,
 					createCommandAction(new SampleScriptLoader(this), "Open sample scripts"),
@@ -801,8 +801,17 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		// Listen for cache request changes
 		PathPrefs.useProjectImageCacheProperty().addListener(v -> updateProjectActionStates());
 		
-		// Menus should now be complete
+		// Menus should now be complete - try binding visibility
 		initializingMenus.set(false);
+		try {
+			for (var item : MenuTools.getFlattenedMenuItems(menuBar.getMenus(), false)) {
+				if (!item.visibleProperty().isBound())
+					bindVisibilityForExperimental(item);
+			}
+		} catch (Exception e) {
+			logger.warn("Error binding menu visibility: {}", e.getLocalizedMessage());
+			logger.warn("", e);
+		}
 		
 		// Update the title
 		stage.titleProperty().bind(titleBinding);
@@ -1362,8 +1371,10 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 	
 	/**
 	 * Populate the availablePathClasses with a default list.
+	 * 
+	 * @return true if changes were mad to the available classes, false otherwise
 	 */
-	public void resetAvailablePathClasses() {
+	public boolean resetAvailablePathClasses() {
 		List<PathClass> pathClasses = new ArrayList<>(); 
 		pathClasses.add(PathClassFactory.getPathClassUnclassified());
 		
@@ -1375,10 +1386,11 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		pathClasses.add(PathClassFactory.getPathClass(PathClassFactory.StandardPathClasses.REGION));
 		pathClasses.add(PathClassFactory.getPathClass(PathClassFactory.StandardPathClasses.IGNORE));
 		
-		if (availablePathClasses == null)
+		if (availablePathClasses == null) {
 			availablePathClasses = FXCollections.observableArrayList(pathClasses);
-		else
-			availablePathClasses.setAll(pathClasses);
+			return true;
+		} else
+			return availablePathClasses.setAll(pathClasses);
 	}
 	
 	/**
@@ -1794,7 +1806,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		viewer.getView().addEventFilter(ScrollEvent.ANY, new ScrollEventPanningFilter(viewer));
 		
 		
-		viewer.getView().addEventFilter(KeyEvent.KEY_PRESSED, e -> {
+		viewer.getView().addEventFilter(KeyEvent.KEY_RELEASED, e -> {
 			if (!e.isConsumed()) {
 				PathObject pathObject = viewer.getSelectedObject();
 				if (pathObject instanceof TMACoreObject) {
@@ -1809,7 +1821,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 					}
 				} else if (pathObject instanceof PathAnnotationObject) {
 					if (e.getCode() == KeyCode.ENTER) {
-						PathAnnotationPanel.promptToSetActiveAnnotationProperties(viewer.getHierarchy());
+						GuiTools.promptToSetActiveAnnotationProperties(viewer.getHierarchy());
 						e.consume();
 					}
 				}
@@ -1947,7 +1959,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		});
 		MenuItem miToggleSync = getActionCheckBoxMenuItem(GUIActions.TOGGLE_SYNCHRONIZE_VIEWERS, null);
 		MenuItem miMatchResolutions = getActionMenuItem(GUIActions.MATCH_VIEWER_RESOLUTIONS);
-		Menu menuMultiview = createMenu(
+		Menu menuMultiview = MenuTools.createMenu(
 				"Multi-view",
 				miToggleSync,
 				miMatchResolutions,
@@ -1964,7 +1976,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				miRemoveColumn
 				);
 		
-		Menu menuView = createMenu(
+		Menu menuView = MenuTools.createMenu(
 				"Display",
 				getActionCheckBoxMenuItem(GUIActions.SHOW_ANALYSIS_PANEL, null),
 				getAction(GUIActions.BRIGHTNESS_CONTRAST),
@@ -1980,7 +1992,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				getActionCheckBoxMenuItem(GUIActions.ZOOM_TO_FIT, null)
 				);
 		
-		Menu menuTools = createMenu(
+		Menu menuTools = MenuTools.createMenu(
 				"Set tool",
 				getActionCheckBoxMenuItem(GUIActions.MOVE_TOOL, null),
 				getActionCheckBoxMenuItem(GUIActions.RECTANGLE_TOOL, null),
@@ -1994,16 +2006,6 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				);
 
 		
-		// Add annotation options
-		Menu menuAnnotations = createMenu(
-				"Annotations",
-				createCommandAction(new AnnotationCombineCommand(viewer, RoiTools.CombineOp.ADD), "Merge selected"),
-				createCommandAction(new AnnotationCombineCommand(viewer, RoiTools.CombineOp.SUBTRACT), "Subtract selected"), // TODO: Make this less ambiguous!
-				createCommandAction(new AnnotationCombineCommand(viewer, RoiTools.CombineOp.INTERSECT), "Intersect selected"),
-				createCommandAction(new InverseObjectCommand(this), "Make inverse"),
-				createPluginAction("Split selected", SplitAnnotationsPlugin.class, null)
-				);
-		
 		// Handle awkward 'TMA core missing' option
 		CheckMenuItem miTMAValid = new CheckMenuItem("Set core valid");
 		miTMAValid.setOnAction(e -> setTMACoreMissing(viewer.getHierarchy(), false));
@@ -2011,21 +2013,21 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		miTMAMissing.setOnAction(e -> setTMACoreMissing(viewer.getHierarchy(), true));
 		
 		Menu menuTMA = new Menu("TMA");
-		addMenuItems(
+		MenuTools.addMenuItems(
 				menuTMA,
 				miTMAValid,
 				miTMAMissing,
 				null,
 				getAction(GUIActions.TMA_ADD_NOTE),
 				null,
-				createMenu(
+				MenuTools.createMenu(
 						"Add",
 					createCommandAction(new TMAGridAdd(this, TMAAddType.ROW_BEFORE), "Add TMA row before"),
 					createCommandAction(new TMAGridAdd(this, TMAAddType.ROW_AFTER), "Add TMA row after"),
 					createCommandAction(new TMAGridAdd(this, TMAAddType.COLUMN_BEFORE), "Add TMA column before"),
 					createCommandAction(new TMAGridAdd(this, TMAAddType.COLUMN_AFTER), "Add TMA column after")
 					),
-				createMenu(
+				MenuTools.createMenu(
 						"Remove",
 					createCommandAction(new TMAGridRemove(this, TMARemoveType.ROW), "Remove TMA row"),
 					createCommandAction(new TMAGridRemove(this, TMARemoveType.COLUMN), "Remove TMA column")
@@ -2035,20 +2037,8 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		
 		
 		// Create an empty placeholder menu
-		Menu menuSetClass = createMenu("Set class");
+		Menu menuSetClass = MenuTools.createMenu("Set class");
 		
-		
-		Action actionInsert = createCommandAction(new HierarchyInsertCommand(this), "Insert in hierarchy");
-		CheckMenuItem miLockAnnotations = new CheckMenuItem("Lock");
-		CheckMenuItem miUnlockAnnotations = new CheckMenuItem("Unlock");
-		miLockAnnotations.setOnAction(e -> setSelectedAnnotationLock(viewer.getHierarchy(), true));
-		miUnlockAnnotations.setOnAction(e -> setSelectedAnnotationLock(viewer.getHierarchy(), false));
-		menuAnnotations.getItems().addAll(0, Arrays.asList(
-				miLockAnnotations,
-				miUnlockAnnotations,
-				new SeparatorMenuItem(),
-				ActionUtils.createMenuItem(actionInsert),
-				new SeparatorMenuItem()));
 		
 //		CheckMenuItem miTMAValid = new CheckMenuItem("Set core valid");
 //		miTMAValid.setOnAction(e -> setTMACoreMissing(viewer.getHierarchy(), false));
@@ -2056,11 +2046,11 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 //		miTMAMissing.setOnAction(e -> setTMACoreMissing(viewer.getHierarchy(), true));
 		
 		
-		Menu menuCells = createMenu(
+		Menu menuCells = MenuTools.createMenu(
 				"Cells",
-				createRadioMenuItem(getAction(GUIActions.SHOW_CELL_BOUNDARIES_AND_NUCLEI), null),
-				createRadioMenuItem(getAction(GUIActions.SHOW_CELL_NUCLEI), null),
-				createRadioMenuItem(getAction(GUIActions.SHOW_CELL_BOUNDARIES), null)
+				MenuTools.createRadioMenuItem(getAction(GUIActions.SHOW_CELL_BOUNDARIES_AND_NUCLEI), null),
+				MenuTools.createRadioMenuItem(getAction(GUIActions.SHOW_CELL_NUCLEI), null),
+				MenuTools.createRadioMenuItem(getAction(GUIActions.SHOW_CELL_BOUNDARIES), null)
 				);
 
 		
@@ -2077,6 +2067,8 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 			}
 		});
 		
+		// Create a standard annotations menu
+		Menu menuAnnotations = GuiTools.populateAnnotationsMenu(this, new Menu("Annotations"));
 		
 		SeparatorMenuItem topSeparator = new SeparatorMenuItem();
 		popup.setOnShowing(e -> {
@@ -2097,10 +2089,6 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				miTMAValid.setSelected(!isMissing);
 				miTMAMissing.setSelected(isMissing);
 				menuTMA.setVisible(true);
-			} else if (pathObject instanceof PathAnnotationObject) {
-				boolean isLocked = ((PathAnnotationObject)pathObject).isLocked();
-				miLockAnnotations.setSelected(isLocked);
-				miUnlockAnnotations.setSelected(!isLocked);
 			}
 			
 			// Add clear objects option if we have more than one non-TMA object
@@ -2166,8 +2154,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 //		});
 			
 	}
-	
-	
+
 	/**
 	 * Set selected TMA cores to have the specified 'missing' status.
 	 * 
@@ -2197,39 +2184,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		if (!changed.isEmpty())
 			hierarchy.fireObjectsChangedEvent(getInstance(), changed);
 	}
-	
-	
-	/**
-	 * Set selected TMA cores to have the specified 'locked' status.
-	 * 
-	 * @param hierarchy
-	 * @param setToLocked
-	 */
-	private static void setSelectedAnnotationLock(final PathObjectHierarchy hierarchy, final boolean setToLocked) {
-		if (hierarchy == null)
-			return;
-		PathObject pathObject = hierarchy.getSelectionModel().getSelectedObject();
-		List<PathObject> changed = new ArrayList<>();
-		if (pathObject instanceof PathAnnotationObject) {
-			PathAnnotationObject annotation = (PathAnnotationObject)pathObject;
-			annotation.setLocked(setToLocked);
-			changed.add(annotation);
-			// Update any other selected cores to have the same status
-			for (PathObject pathObject2 : hierarchy.getSelectionModel().getSelectedObjects()) {
-				if (pathObject2 instanceof PathAnnotationObject) {
-					annotation = (PathAnnotationObject)pathObject2;
-					if (annotation.isLocked() != setToLocked) {
-						annotation.setLocked(setToLocked);
-						changed.add(annotation);
-					}
-				}
-			}
-		}
-		if (!changed.isEmpty())
-			hierarchy.fireObjectsChangedEvent(getInstance(), changed);
-	}
-	
-	
+		
 	
 	/**
 	 * Update a 'set annotation class' menu for a viewer immediately prior to display
@@ -2999,59 +2954,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		return menuBar;
 	}
 	
-	/**
-	 * Create a menu, and add new menu items.
-	 * 
-	 * If null is passed as an object, a separator is added.
-	 * 
-	 * @param name
-	 * @param items
-	 * @return new menu
-	 */
-	public static Menu createMenu(final String name, final Object... items) {
-		return addMenuItems(createMenu(name), items);
-	}
 	
-	/**
-	 * Add menu items to an existing menu.
-	 * 
-	 * If null is passed as an object, a separator is added.
-	 * 
-	 * @param menu
-	 * @param items
-	 * @return menu, so that this method can be nested inside other calls.
-	 */
-	public static Menu addMenuItems(final Menu menu, final Object... items) {
-		// Check if the last item was a separator -
-		// we don't want two adjacent separators, since this looks a bit weird
-		boolean lastIsSeparator = menu.getItems().isEmpty() ? false : menu.getItems().get(menu.getItems().size()-1) instanceof SeparatorMenuItem;
-		
-		List<MenuItem> newItems = new ArrayList<>();
-		for (Object item : items) {
-			if (item == null) {
-				if (!lastIsSeparator)
-					newItems.add(new SeparatorMenuItem());
-				lastIsSeparator = true;
-			}
-			else if (item instanceof MenuItem) {
-				newItems.add((MenuItem)item);
-				lastIsSeparator = false;
-			}
-			else if (item instanceof Action) {
-				newItems.add(createMenuItem((Action)item));
-				lastIsSeparator = false;
-			} else
-				logger.warn("Could not add menu item {}", item);
-		}
-		if (!newItems.isEmpty()) {
-			boolean initializing = initializingMenus.get();
-			initializingMenus.set(true);
-//			System.err.println("MENU STUFF: " + initializingMenus.get() + " - " + menu.getText());
-			menu.getItems().addAll(newItems);
-			initializingMenus.set(initializing);
-		}
-		return menu;
-	}
 	
 	
 	/**
@@ -3116,17 +3019,19 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 
 	
 	
-	protected MenuBar createMenuBar() {
+	private MenuBar createMenuBar() {
 		
 		// Create a recent projects list
 		ObservableList<URI> recentProjects = PathPrefs.getRecentProjectList();
-		Menu menuRecent = createMenu("Recent projects...");
+		Menu menuRecent = MenuTools.createMenu("Recent projects...");
 		
 		
 		// Create a File menu
-		Menu menuFile = createMenu(
+		Action actionQuit = new Action("Quit", e -> tryToQuit());
+//		actionQuit.setAccelerator(new KeyCodeCombination(KeyCode.Q, KeyCodeCombination.SHORTCUT_DOWN));
+		Menu menuFile = MenuTools.createMenu(
 				"File",
-				createMenu(
+				MenuTools.createMenu(
 						"Project...",
 						getActionMenuItem(GUIActions.PROJECT_NEW),
 						getActionMenuItem(GUIActions.PROJECT_OPEN),
@@ -3147,24 +3052,26 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				getActionMenuItem(GUIActions.SAVE_DATA_AS),
 				getActionMenuItem(GUIActions.SAVE_DATA),
 				null,
-				createMenu(
+				MenuTools.createMenu(
 						"Export images...",
 						createCommandAction(new ExportImageRegionCommand(this, false), "Original pixels"),
 						createCommandAction(new ExportImageRegionCommand(this, true), "Rendered RGB (with overlays)")
 						),
-				createMenu(
+				MenuTools.createMenu(
 						"Export snapshot...",
 						createCommandAction(new SaveViewCommand(this, GuiTools.SnapshotType.MAIN_WINDOW_SCREENSHOT), "Main window screenshot"),
 						createCommandAction(new SaveViewCommand(this, GuiTools.SnapshotType.MAIN_SCENE), "Main window content"),
 						createCommandAction(new SaveViewCommand(this, GuiTools.SnapshotType.VIEWER), "Current viewer content")
 						),
 				null,
-				createMenu(
+				MenuTools.createMenu(
 						"TMA data...",
 						getActionMenuItem(GUIActions.TMA_SCORE_IMPORTER),
 						getActionMenuItem(GUIActions.TMA_EXPORT_DATA),
 						createCommandAction(new TMAViewerCommand(), "Launch TMA data viewer")
-						)
+						),
+				null,
+				actionQuit
 				);
 		
 		
@@ -3196,12 +3103,12 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		
 		
 		// Create Edit menu
-		Menu menuEdit = createMenu(
+		Menu menuEdit = MenuTools.createMenu(
 				"Edit",
 				getActionMenuItem(GUIActions.UNDO),
 				getActionMenuItem(GUIActions.REDO),
 				null,
-				createMenu("Copy to clipboard...",
+				MenuTools.createMenu("Copy to clipboard...",
 					getActionMenuItem(GUIActions.COPY_VIEW),
 					getActionMenuItem(GUIActions.COPY_WINDOW),
 					getActionMenuItem(GUIActions.COPY_WINDOW_SCREENSHOT),
@@ -3214,7 +3121,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 
 		// Create Tools menu
 		ToggleGroup groupTools = new ToggleGroup();
-		Menu menuTools = createMenu(
+		Menu menuTools = MenuTools.createMenu(
 				"Tools",
 				getActionCheckBoxMenuItem(GUIActions.MOVE_TOOL, groupTools),
 				getActionCheckBoxMenuItem(GUIActions.RECTANGLE_TOOL, groupTools),
@@ -3227,8 +3134,8 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				getActionCheckBoxMenuItem(GUIActions.POINTS_TOOL, groupTools)
 				);
 		
-		Menu menuGestures = createMenu("Multi-touch gestures");
-		addMenuItems(
+		Menu menuGestures = MenuTools.createMenu("Multi-touch gestures");
+		MenuTools.addMenuItems(
 				menuGestures,
 				ActionUtils.createMenuItem(new Action("Turn on all gestures", e -> {
 					PathPrefs.setUseScrollGestures(true);
@@ -3245,7 +3152,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				ActionUtils.createCheckMenuItem(createSelectableCommandAction(PathPrefs.useZoomGesturesProperty(), "Use zoom gestures")),
 				ActionUtils.createCheckMenuItem(createSelectableCommandAction(PathPrefs.useRotateGesturesProperty(), "Use rotate gestures"))
 				);
-		addMenuItems(
+		MenuTools.addMenuItems(
 				menuTools,
 				null,
 				menuGestures);
@@ -3270,7 +3177,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		// Create View menu
 		SlideLabelView slideLabelView = new SlideLabelView(this);
 		ToggleGroup groupCellDisplay = new ToggleGroup();
-		Menu menuView = createMenu(
+		Menu menuView = MenuTools.createMenu(
 				"View",
 				getActionCheckBoxMenuItem(GUIActions.SHOW_ANALYSIS_PANEL),
 				getActionMenuItem(GUIActions.SHOW_COMMAND_LIST),
@@ -3280,13 +3187,13 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				null,
 				getActionCheckBoxMenuItem(GUIActions.TOGGLE_SYNCHRONIZE_VIEWERS),
 				getActionCheckBoxMenuItem(GUIActions.MATCH_VIEWER_RESOLUTIONS),
-				createMenu(
+				MenuTools.createMenu(
 						"Mini viewers",
 						getActionMenuItem(GUIActions.CHANNEL_VIEWER),
 						getActionMenuItem(GUIActions.MINI_VIEWER)
 						),
 				null,
-				createMenu(
+				MenuTools.createMenu(
 						"Zoom",
 						createCommandAction(new ViewerSetDownsampleCommand(this, 0.25), "400%"),
 						createCommandAction(new ViewerSetDownsampleCommand(this, 1), "100%"),
@@ -3296,11 +3203,11 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 						null,
 						getActionMenuItem(GUIActions.ZOOM_IN),
 						getActionMenuItem(GUIActions.ZOOM_OUT),
-						createCheckMenuItem(getAction(GUIActions.ZOOM_TO_FIT))
+						MenuTools.createCheckMenuItem(getAction(GUIActions.ZOOM_TO_FIT))
 						),
 				getActionMenuItem(GUIActions.ROTATE_IMAGE),
 				null,
-				createMenu(
+				MenuTools.createMenu(
 						"Cell display",
 						getActionCheckBoxMenuItem(GUIActions.SHOW_CELL_BOUNDARIES, groupCellDisplay),
 						getActionCheckBoxMenuItem(GUIActions.SHOW_CELL_NUCLEI, groupCellDisplay),
@@ -3312,7 +3219,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				getActionCheckBoxMenuItem(GUIActions.SHOW_TMA_GRID_LABELS),
 				getActionCheckBoxMenuItem(GUIActions.SHOW_DETECTIONS),
 				getActionCheckBoxMenuItem(GUIActions.FILL_DETECTIONS),
-				createCheckMenuItem(createSelectableCommandAction(overlayOptions.showConnectionsProperty(), "Show object connections")),
+				MenuTools.createCheckMenuItem(createSelectableCommandAction(overlayOptions.showConnectionsProperty(), "Show object connections")),
 				getActionCheckBoxMenuItem(GUIActions.SHOW_PIXEL_CLASSIFICATION),
 				null,
 				getActionCheckBoxMenuItem(GUIActions.SHOW_OVERVIEW),
@@ -3322,7 +3229,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				getActionMenuItem(GUIActions.GRID_SPACING),
 				null,
 				getActionMenuItem(GUIActions.VIEW_TRACKER),
-				createCheckMenuItem(createSelectableCommandAction(slideLabelView.showingProperty(), "Show slide label")),				
+				MenuTools.createCheckMenuItem(createSelectableCommandAction(slideLabelView.showingProperty(), "Show slide label")),				
 				null,
 				createCommandAction(new ShowInputDisplayCommand(this), "Show input on screen"),
 				createCommandAction(new MemoryMonitorCommand(this), "Show memory monitor"),
@@ -3330,9 +3237,9 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 			);
 		
 		
-		Menu menuObjects = createMenu(
+		Menu menuObjects = MenuTools.createMenu(
 				"Objects",
-				createMenu(
+				MenuTools.createMenu(
 						"Delete...",
 						getActionMenuItem(GUIActions.DELETE_SELECTED_OBJECTS),
 						null,
@@ -3340,7 +3247,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 						getActionMenuItem(GUIActions.CLEAR_ANNOTATIONS),
 						getActionMenuItem(GUIActions.CLEAR_DETECTIONS)
 						),
-				createMenu(
+				MenuTools.createMenu(
 						"Select...",
 						createCommandAction(new ResetSelectionCommand(this), "Reset selection", null, new KeyCodeCombination(KeyCode.R, KeyCodeCombination.SHORTCUT_DOWN, KeyCodeCombination.ALT_DOWN)),
 						null,
@@ -3352,7 +3259,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 						createCommandAction(new SelectObjectsByMeasurementCommand(this), "Select by measurements (experimental)")
 						),
 				null,
-				createMenu("Annotations...",
+				MenuTools.createMenu("Annotations...",
 					getActionMenuItem(GUIActions.SPECIFY_ANNOTATION),
 					getActionMenuItem(GUIActions.SELECT_ALL_ANNOTATION),
 					null,					
@@ -3375,16 +3282,16 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				);
 
 		
-		Menu menuTMA = createMenu(
+		Menu menuTMA = MenuTools.createMenu(
 				"TMA",
-				createMenu(
+				MenuTools.createMenu(
 					"Add...",
 					createCommandAction(new TMAGridAdd(this, TMAAddType.ROW_BEFORE), "Add TMA row before"),
 					createCommandAction(new TMAGridAdd(this, TMAAddType.ROW_AFTER), "Add TMA row after"),
 					createCommandAction(new TMAGridAdd(this, TMAAddType.COLUMN_BEFORE), "Add TMA column before"),
 					createCommandAction(new TMAGridAdd(this, TMAAddType.COLUMN_AFTER), "Add TMA column after")
 					),
-				createMenu(
+				MenuTools.createMenu(
 						"Remove...",
 						createCommandAction(new TMAGridRemove(this, TMARemoveType.ROW), "Remove TMA row"),
 						createCommandAction(new TMAGridRemove(this, TMARemoveType.COLUMN), "Remove TMA column")
@@ -3399,7 +3306,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				);
 		
 		
-		Menu menuMeasure = createMenu(
+		Menu menuMeasure = MenuTools.createMenu(
 				"Measure",
 				getActionMenuItem(GUIActions.MEASUREMENT_MAP),
 				createCommandAction(new MeasurementManager(this), "Show measurement manager"),
@@ -3417,7 +3324,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		// Try to load a script editor
 		Action actionScriptEditor = createCommandAction(new ShowScriptEditorCommand(this, false), "Show script editor");
 		actionScriptEditor.setAccelerator(new KeyCodeCombination(KeyCode.OPEN_BRACKET, KeyCodeCombination.SHORTCUT_DOWN, KeyCodeCombination.SHIFT_ANY));
-		Menu menuAutomate = createMenu(
+		Menu menuAutomate = MenuTools.createMenu(
 				"Automate",
 				actionScriptEditor,
 				createCommandAction(new ScriptInterpreterCommand(this), "Script interpreter"),
@@ -3427,23 +3334,23 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				);
 		
 		// Add some plugins
-		Menu menuAnalysis = createMenu(
+		Menu menuAnalysis = MenuTools.createMenu(
 				"Analyze",
-				createMenu(
+				MenuTools.createMenu(
 						"Preprocessing",
 						getActionMenuItem(GUIActions.COLOR_DECONVOLUTION_REFINE)
 						),
-				createMenu(
+				MenuTools.createMenu(
 						"Region identification",
-						createMenu(
+						MenuTools.createMenu(
 								"Tiles & superpixels",
 								createPluginAction("Create tiles", TilerPlugin.class, this, null)
 								)
 						),
-				createMenu(
+				MenuTools.createMenu(
 						"Cell detection"
 						),
-				createMenu(
+				MenuTools.createMenu(
 						"Calculate features",
 //						new PathPluginAction("Create tiles", TilerPlugin.class, this),
 						createPluginAction("Add Smoothed features", SmoothFeaturesPlugin.class, this, null),
@@ -3455,7 +3362,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 //						null,
 //						createPluginAction("Add Local Binary Pattern features (experimental)", LocalBinaryPatternsPlugin.class, this, null)
 						),
-				createMenu(
+				MenuTools.createMenu(
 						"Spatial analysis",
 						createCommandAction(new DistanceToAnnotationsCommand(this), "Distance to annotations 2D (experimental)"),
 						createCommandAction(new DetectionCentroidDistancesCommand(this), "Detection centroid distances 2D (experimental)")
@@ -3463,13 +3370,13 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				);
 
 		// Try to load classifiers
-		Menu menuClassifiers = createMenu("Classify");
+		Menu menuClassifiers = MenuTools.createMenu("Classify");
 
-		Menu menuObjectClassifiers = createMenu("Object classification");
-		Menu menuPixelClassifiers = createMenu("Pixel classification");
-		addMenuItems(menuClassifiers, menuObjectClassifiers, menuPixelClassifiers);
+		Menu menuObjectClassifiers = MenuTools.createMenu("Object classification");
+		Menu menuPixelClassifiers = MenuTools.createMenu("Pixel classification");
+		MenuTools.addMenuItems(menuClassifiers, menuObjectClassifiers, menuPixelClassifiers);
 		
-		addMenuItems(
+		MenuTools.addMenuItems(
 				menuObjectClassifiers,
 				createCommandAction(new LoadClassifierCommand(this), "Load detection classifier"),
 				null,
@@ -3479,7 +3386,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 				createCommandAction(new SingleFeatureClassifierCommand(this, PathDetectionObject.class), "Classify by specific feature")
 				);
 				
-		addMenuItems(
+		MenuTools.addMenuItems(
 				menuClassifiers,
 				null,
 				createCommandAction(new SparseImageServerCommand(this), "Create project training image")
@@ -3489,7 +3396,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 			checkForUpdate(false);
 		});
 		
-		Menu menuHelp = createMenu(
+		Menu menuHelp = MenuTools.createMenu(
 				"Help",
 //				createCommandAction(new HelpCommand(this), "Documentation"),
 				getAction(GUIActions.QUPATH_SETUP),
@@ -3567,6 +3474,23 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		return action;
 	}
 	
+	
+	/**
+	 * Request to quit QuPath.
+	 */
+	void tryToQuit() {
+		var stage = getStage();
+		if (stage == null || !stage.isShowing())
+			return;
+		stage.fireEvent(
+				new WindowEvent(
+				        stage,
+				        WindowEvent.WINDOW_CLOSE_REQUEST
+				    )
+				);
+	}
+	
+	
 	/**
 	 * Create an Action to construct and run a plugin interactively.
 	 * 
@@ -3639,51 +3563,12 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 	}
 	
 	
-	
-//	public static Action createPluginAction(final String name, final Class<? extends PathPlugin> pluginClass, final QuPathGUI qupath, final DefaultImageRegionStore regionStore, final String arg) {
-//		Action action = new Action(name, event -> {
-//		
-//			QuPathViewer viewer = qupath.getViewer();
-//			ImageServer<BufferedImage> server = viewer.getServer();
-//			if (server == null) {
-//				// TODO: Display an error message to the user!
-//				logger.error("No whole slide image server could be found!");
-//				return;
-//			}
-//			try {
-//				PathPlugin plugin;
-//				if (regionStore != null)
-//					plugin = pluginClass.getConstructor(ImageRegionStore.class).newInstance(regionStore);
-//				else
-//					plugin = pluginClass.getConstructor().newInstance();
-//
-//				// TODO: Check safety...
-//				if (plugin instanceof PathInteractivePlugin)
-//					((PathInteractivePlugin<BufferedImage>)plugin).runInteractive(PluginRunnerFactory.makePluginRunnerFX(qupath, false), arg);
-//				else
-//					((PathPlugin<BufferedImage>)plugin).runPlugin(PluginRunnerFactory.makePluginRunnerFX(qupath, false), arg);
-//							
-//			} catch (Exception e) {
-//				logger.error("Unable to initialize class " + pluginClass, e);
-//			}
-//			
-//		});
-//		
-//		return action;
-//	}
-	
-	
-	
 	public static Action createCommandAction(final PathCommand command, final String name, final Node icon, final KeyCombination accelerator) {
 		Action action = new Action(name, e -> command.run());
 		action.setAccelerator(accelerator);
 		action.setGraphic(icon);
 		return action;
 	}
-	
-//	Action createCommandAction(final PathCommand command, final String name, final PathIconFactory.PathIcons icon, final KeyCombination accelerator) {
-//		return createCommandAction(command, name, PathIconFactory.createNode(iconSize, iconSize, icon), accelerator);
-//	}
 
 	public Action createCommandAction(final String className, final String name, final Object... arguments) {
 		try {
@@ -3722,11 +3607,11 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 //		return new PathSelectableAction(command, name, node, accelerator);
 	}
 	
-	private Action createSelectableCommandAction(final ObservableValue<Boolean> property, final String name, final PathIconFactory.PathIcons icon, final KeyCombination accelerator) {
+	private static Action createSelectableCommandAction(final ObservableValue<Boolean> property, final String name, final PathIconFactory.PathIcons icon, final KeyCombination accelerator) {
 		return createSelectableCommandAction(property, name, PathIconFactory.createNode(iconSize, iconSize, icon), accelerator);
 	}
 
-	private Action createSelectableCommandAction(final PathSelectableCommand command, final String name, final Node icon, final KeyCombination accelerator) {
+	private static Action createSelectableCommandAction(final PathSelectableCommand command, final String name, final Node icon, final KeyCombination accelerator) {
 		Action action = new Action(name, e -> command.setSelected(!command.isSelected()));
 		action.selectedProperty().addListener(e -> {
 			command.setSelected(action.isSelected());
@@ -3760,14 +3645,10 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		toolbar.toolbar.getItems().add(button);
 	}
 	
-//	public void addToolbarCommand(final String name, final PathCommand command, final Icon icon) {
-//		toolbar.toolbar.getItems().add(getActionButton(createCommandAction(command, name, PathIconFactory.createNode(icon), null), icon != null));
-//	}
-	
 
 	
 	
-	protected Action createAction(GUIActions actionType) {
+	private Action createAction(GUIActions actionType) {
 //		QuPathViewerPlus viewer = viewerManager.getActiveViewer();
 		Action action;
 		switch (actionType) {
@@ -3936,6 +3817,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		
 		case DELETE_SELECTED_OBJECTS:
 			return createCommandAction(new DeleteSelectedObjectsCommand(this), "Delete selected objects");
+//			return createCommandAction(new DeleteSelectedObjectsCommand(this), "Delete selected objects", null, new KeyCodeCombination(KeyCode.BACK_SPACE));
 		case CLEAR_HIERARCHY:
 			return createCommandAction(new DeleteObjectsCommand(this, null), "Delete all objects");
 		case CLEAR_DETECTIONS:
@@ -4155,7 +4037,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 	
 	public MenuItem getActionMenuItem(GUIActions actionType) {
 		Action action = getAction(actionType);
-		return createMenuItem(action);
+		return MenuTools.createMenuItem(action);
 	}
 	
 	public static CheckBox createCheckBox(final Action action) {
@@ -4163,13 +4045,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 	}
 	
 	
-	public static MenuItem createMenuItem(final Action action) {
-		return bindVisibilityForExperimental(ActionUtils.createMenuItem(action));
-	}
 	
-	public static CheckMenuItem createCheckMenuItem(final Action action) {
-		return bindVisibilityForExperimental(ActionUtils.createCheckMenuItem(action));
-	}
 	
 	/**
 	 * Flag to indicate that menus are being initialized.
@@ -4184,7 +4060,10 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 	private static BooleanBinding showLegacyOptions = PathPrefs.showLegacyOptionsProperty().or(initializingMenus);
 	
 	private static <T extends MenuItem> T bindVisibilityForExperimental(final T menuItem) {
-		String text = menuItem.getText().toLowerCase().trim();
+		String text = menuItem.getText();
+		if (text == null)
+			return menuItem;
+		text = text.toLowerCase().trim();
 		if (text.equals("experimental") || text.endsWith("experimental)"))
 			menuItem.visibleProperty().bind(showExperimentalOptions);
 		else if (text.equals("tma") || text.endsWith("tma)"))
@@ -4195,22 +4074,16 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 	}
 	
 	
-	private static RadioMenuItem createRadioMenuItem(final Action action, final ToggleGroup group) {
-		RadioMenuItem menuItem = ActionUtils.createRadioMenuItem(action);
-		menuItem.setToggleGroup(group);
-		return bindVisibilityForExperimental(menuItem);
-	}
 	
-
-	public MenuItem getActionCheckBoxMenuItem(GUIActions actionType, ToggleGroup group) {
+	private MenuItem getActionCheckBoxMenuItem(GUIActions actionType, ToggleGroup group) {
 		Action action = getAction(actionType);
 		if (group != null)
-			return createRadioMenuItem(action, group);
+			return MenuTools.createRadioMenuItem(action, group);
 		else
-			return createCheckMenuItem(action);
+			return MenuTools.createCheckMenuItem(action);
 	}
 	
-	public MenuItem getActionCheckBoxMenuItem(GUIActions actionType) {
+	private MenuItem getActionCheckBoxMenuItem(GUIActions actionType) {
 		return getActionCheckBoxMenuItem(actionType, null);
 	}
 	
@@ -4226,7 +4099,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		return button;
 	}
 	
-	public ToggleButton getActionToggleButton(GUIActions actionType, boolean hideActionText, ToggleGroup group) {
+	private ToggleButton getActionToggleButton(GUIActions actionType, boolean hideActionText, ToggleGroup group) {
 		Action action = getAction(actionType);
 		ToggleButton button = ActionUtils.createToggleButton(action, hideActionText ? ActionTextBehavior.HIDE : ActionTextBehavior.SHOW);
 		if (hideActionText && action.getText() != null) {
@@ -4255,7 +4128,7 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 		return button;
 	}
 
-	public ToggleButton getActionToggleButton(GUIActions actionType, boolean hideActionText, boolean isSelected) {
+	private ToggleButton getActionToggleButton(GUIActions actionType, boolean hideActionText, boolean isSelected) {
 		return getActionToggleButton(actionType, hideActionText, null, isSelected);
 	}
 	
@@ -4362,96 +4235,32 @@ public class QuPathGUI implements ModeWrapper, ImageDataWrapper<BufferedImage>, 
 //		viewer.centerImage();
 //	}
 	
-	
+	/**
+	 * Get the menubar for the main QuPath application.
+	 * @return
+	 */
 	public MenuBar getMenuBar() {
 		return menuBar;
 	}
 
 	/**
-	 * Get a reference to an existing menu, optionally creating a new menu if it is not present.
+	 * Get a reference to an existing menu from the main QuPath menubar, optionally creating a new menu if it is not present.
 	 * 
 	 * @param name
 	 * @param createMenu
 	 * @return
 	 */
 	public Menu getMenu(final String name, final boolean createMenu) {
-		MenuBar menuBar = getMenuBar();
-		if (menuBar == null)
+		var menubar = getMenuBar();
+		if (menubar == null)
 			return null;
-		
-		Menu menuCurrent = null;
-		for (String n : name.split(">")) {
-			if (menuCurrent == null) {
-				for (Menu menu : menuBar.getMenus()) {
-					if (n.equals(menu.getText())) {
-						menuCurrent = menu;
-						break;
-					}
-				}
-				if (menuCurrent == null) {
-					if (createMenu) {
-						menuCurrent = new Menu(n.trim());
-						// Make sure we don't replace the 'Help' menu at the end
-						List<Menu> menus = menuBar.getMenus();
-						if (!menus.isEmpty() && "Help".equals(menus.get(menus.size()-1).getText()))
-							menus.add(menus.size()-1, menuCurrent);
-						else
-							menus.add(menuCurrent);
-					} else
-						return null;
-				}
-			} else {
-				List<MenuItem> searchItems = menuCurrent.getItems();
-				menuCurrent = null;
-				for (MenuItem menuItem : searchItems) {
-					if (menuItem instanceof Menu && (menuItem.getText().equals(n) || menuItem.getText().equals(n.trim()))) {
-						menuCurrent = (Menu)menuItem;
-						break;
-					}
-				}
-				if (menuCurrent == null) {
-					if (createMenu) {
-						menuCurrent = new Menu(n.trim());
-						searchItems.add(menuCurrent);
-					} else
-						return null;
-				}				
-			}
-		}
-		return menuCurrent;
-	}
-	
-	
-	public MenuItem getMenuItem(String itemName) {
-		Collection<MenuItem> menuItems;
-		int ind = itemName.lastIndexOf(">");
-		if (ind >= 0) {
-			Menu menu = getMenu(itemName.substring(0, ind), false);
-			if (menu == null) {
-				logger.warn("No menu found for {}", itemName);
-				return null;
-			}
-			menuItems = menu.getItems();
-			itemName = itemName.substring(ind+1);
-		} else {
-			menuItems = new HashSet<>();
-			for (Menu menu : getMenuBar().getMenus())
-				menuItems.addAll(menu.getItems());
-		}
-		for (MenuItem menuItem : menuItems) {
-			if (itemName.equals(menuItem.getText()))
-				return menuItem;
-		}
-		logger.warn("No menu item found for {}", itemName);
-		return null;
-	}
-	
-	
-	static Menu createMenu(final String name) {
-		return bindVisibilityForExperimental(new Menu(name));
+		return MenuTools.getMenu(menuBar.getMenus(), name, createMenu);
 	}
 
-	
+	/**
+	 * Get the main QuPath stage.
+	 * @return
+	 */
 	public Stage getStage() {
 		return stage;
 	}