+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## Version 0.2.0-m9
+This is a *milestone* (i.e. still in development) version made available to try out new features early.
+* Major changes to annotation ROI manipulation
+  * 'Duplicate annotations' applies to multiple selections
+  * 'Merge annotations' and 'Split annotations' work with point ROIs, not only areas
+  * 'Make inverse' uses ROIs from multiple annotations (within the same plane)
+  * More ROI manipulation commands are scriptable, update selections when complete
+* Revised "Annotations" tab
+  * New options to set available class list
+* File -> Quit menu item added
+* New preferences
+  * Select main font; default changed to Sans-Serif for macOS
+  * Turn on/off system menubar
+* Show accelerator within 'Command list' table
+* Fixed issue with 'Import images' where the window could be too large for some screens
+* Update dependencies (Bio-Formats, ImageJ, ControlsFX, RichTextFX)
+
+
 ## Version 0.2.0-m8
 This is a *milestone* (i.e. still in development) version made available to try out new features early.
 * Fixed repainting bug that could cause existing annotations to temporarily shift when drawing new annotations